🚚 Move configuration, tokens, project, and events view into admin app
