+++ src/admin/views/index.js
@@ -1,0 +1,4 @@
+export {default as CavaticaProjectsView} from './CavaticaProjectsView';
+export {default as ConfigurationView} from './ConfigurationView';
+export {default as EventsView} from './EventsView';
+export {default as TokensListView} from './TokensListView';
