+++ src/views/index.js
@@ -4,15 +4,11 @@ export {default as LogoutView} from './LogoutView';
 export {default as CallbackView} from './CallbackView';
 export {default as NavBarView} from './NavBarView';
 export {default as EmptyView} from './EmptyView';
-export {default as TokensListView} from './TokensListView';
-export {default as CavaticaProjectsView} from './CavaticaProjectsView';
 export {default as NewStudyView} from './NewStudyView';
 export {default as ProfileView} from './ProfileView';
 export {default as StudyInfoView} from './StudyInfoView';
-export {default as EventsView} from './EventsView';
 export {default as CavaticaBixView} from './CavaticaBixView';
 export {default as LogsView} from './LogsView';
-export {default as ConfigurationView} from './ConfigurationView';
 export {default as ReleasesView} from './ReleasesView';
 export {default as NewStudySelectionView} from './NewStudySelectionView';
 export {default as NotFoundView} from './NotFoundView';
